Fix: Resolve VerifyError due to obscure WrapWithCondition/ModifyReceiver + Redirect interaction.
A WrapWithCondition or a ModifyReceiver which captures its target args when applied to a Redirect which doesn't but has sugar would mistakenly load the sugar params instead of the args.

The handlerArgMap needed to be trimmed to remove trailing args before adding the enclosing method's params.

Notably, this interaction could also occur when the applied Redirect captures the enclosing params rather than has sugar, but in this case the interaction is coincidentally benign, with the corresponding inner captured params being loaded instead of the original params.